Cleaned up Prototype folder, created template
All prototype demos should be contained inside of their own folder, and
utilize the structure provided in the _tmpl folder. Use index.html as
the demo page inside a properly named folder. if you have multiple
examples of a prototype, utilize the sub-nav and stick to index.html
being the main demo and then the second demo should be aptly named
according to what the name is.