Using a different bookmarklet. one that works
It's really hacky, but it gets the job done. I recommend that we go with
a button-approach where they load up the bookmarklet themselves. thereis
just too much involved with this to pull it in and expect it to work
correctly IMO. I spent a few hours on this today with a hacky,
non-standard, breaks if you refresh the browser window, approach.